GUI fixes BD security improved Added new parameter Theme in settings.ini Minor project refactoring
